requirements.txt
duckdb
pandas
matplotlib
seaborn
scipy
numpy
PyYAML
Jinja2
